requirements.txt
torch>=2.0
torchvision
gradio
numpy
opencv-python
Pillow
reportlab
matplotlib
transformers